+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix load screen transition clipping when passing through door Styx -> [Redacted] for 32-bit engine.
+
 ## [v1.4.1] - 2021-09-15
 
 ### Fixed
@@ -93,7 +97,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   Fix load screen transition clipping when passing through door Styx -> [Redacted].
+-   Fix load screen transition clipping when passing through door Styx -> [Redacted] for DirectX 64-bit and Vulkan engines.
 -   Fix biome map first time reward icon scaling.
 
 ## [v1.0.0] - 2021-09-01