+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.4.2] - 2021-09-18
+
 ### Changed
 
 -   Improve Hades directory auto-detection on Windows using OS variables (drive- and architecture-agnostic).
@@ -241,7 +243,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Initial release.
 
-[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.4.1...HEAD
+[Unreleased]: https://github.com/nbusseneau/hephaistos/compare/v1.4.2...HEAD
+
+[v1.4.2]: https://github.com/nbusseneau/hephaistos/compare/v1.4.1...v1.4.2
 
 [v1.4.1]: https://github.com/nbusseneau/hephaistos/compare/v1.4.0...v1.4.1